requirements.txt
fastapi
uvicorn
cloudinary
google-generativeai
python-dotenv
requests
python-multipart